throttler: Split /throttlerz/<name>/log page out into /throttlerlogz/<name>.
This way, throttlers with a "/" in their name are supported.
